[MRH/SB][81377330] major version bump
This change is not backwards compatible.
Prior to this change, the code depended on a mixin from AG
to get listing media data.
This version sends an event "uiWantsListingMedia"
with a listingId when it needs media.
It listens for an event "dataListingMediaReady" which should
have the keys "photos" and "floorplans" where each is expected
to be an array of hashes with image "path".